+++ CHANGES.rst
@@ -1,4 +1,17 @@
 Changelog for Solgema.FlowView
+------------------------------
+
+1.0.5
+-----
+- Use plone.app.contenttypes' FolderView as base for FlowView view as we use "folder_listing" in the template.
+
+
+1.0.4
+-----
+
+- Fix getting portal catalog on brain item, use getSiteUI instead
+
+- Fix errors when Martronic.ContentImageExtend is not installed
 
 - 1.0.3 : Fix Plone5 collection. added field for navigation buttons custom class